refactor: add template preference component
